BC-1686 remove /partner
