+++ controllers/index.js
@@ -26,7 +26,6 @@ router.use('/helpdesk/', require('./helpdesk'));
 router.use('/pwrecovery/', require('./pwrecovery'));
 router.use('/notification/', require('./notification'));
 router.use('/link/', require('./link'));
-router.use('/partner/', require('./partner'));
 router.use('/community/', require('./community'));
 router.use('/about/', require('./about'));
 router.use('/help/', require('./help'));
