+++ locales/uk.json
@@ -2524,21 +2524,6 @@
 			"sessionExtended": "Сеанс успішно продовжено."
 		}
 	},
-	"partner": {
-		"headline": {
-			"advisoryBoard": "Консультативна рада",
-			"pilotPartners": "Пілотні партнери для вивчення контенту",
-			"supporters": "Прихильників шкільної інфраструктури < 3"
-		},
-		"text": {
-			"advisoryBoardDescription": "Проєкт шкільної хмари HPI підтримується консультативною радою, до складу якої входять різні представники освітньої галузі, бізнесу та політики. Окрім технічних аспектів, увага також приділяється таким темам, як конфіденційність та педагогіка.",
-			"pilotPartnersDescription": "Ми хотіли б подякувати нашим партнерам, які зробили різноманітний навчальний контент доступним безкоштовно під час пілотної фази, щоб можна було використовувати шкільну хмару HPI.",
-			"supportersDescription": "Дякуємо нашим прихильникам за шкільну інфраструктуру."
-		},
-		"img_alt": {
-			"logo": "Логотип"
-		}
-	},
 	"help": {
 		"search": {
 			"headline": {
