--- static/styles/partner/partner.scss
@@ -1,77 +1,0 @@
-@import '../lib/colors';
-@import '../homepage/layouts';
-
-hgroup {
-    text-align: center;
-    font-weight: bold;
-    margin: 60px 0 60px;
-    width: 100%;
-}
-
-.section-title {
-    width: 100%;
-    line-height: 1.75rem;
-    h4 {
-        color: $primaryColor;
-        margin-bottom: 0;
-        margin-right: 20px;
-        font-weight: 700;
-    }
-}
-
-.section-partner {
-    ul {
-        color: $primaryColor;
-        padding-left: 20px;
-    }
-    .container {
-        margin-left: 0;
-    }
-    .intro-box {
-        padding: 0;
-    }
-    .partner-box-wrapper {
-        padding-top: 15px;
-        padding-bottom: 15px;
-    }
-    .partner-box {
-        background: #fff;
-        border: 1px solid lightgray;
-        height: 150px;
-		text-align: center;
-		&:hover {
-			box-shadow: var(--shadow-sm);
-			transition: box-shadow .2s ease,transform .2s ease,z-index 0s 0s ease;
-        }
-    }
-    .partner-box img {
-        max-height: 100%;
-        /* Vertical Center */
-        position: relative;
-        top: 50%;
-        transform: translateY(-50%);
-    }
-    .collapse.collapse.in {
-        display: flex !important;
-    }
-    .toggle-partner {
-        position: relative;
-        i {
-            display: block;
-            pointer-events: none;
-        }
-    }
-    .transparent{
-        background-color: transparent;
-    }
-    .buttonWrapper {
-        width: 100%;
-        text-align: center;
-    }
-}
-
-.primary, .secondary{
-	.partner-box{
-		border: 1px solid #fff;
-	}
-}
