--- theme/thr/views/partner/partner_guests.hbs
@@ -1,15 +1,0 @@
-{{#extend "lib/loggedout"}}
-
-    {{#content "scripts" mode="append"}}
-        <script src="{{getAssetPath '/scripts/partners.js'}}" type="text/javascript" nonce="{{nonceValue}}" defer></script>
-    {{/content}}
-
-    {{#content "styles" mode="append"}}
-        <link rel="stylesheet" href="{{getAssetPath '/styles/partner/partner.css'}}"/>
-    {{/content}}
-
-    {{#content "page"}}
-		{{#embed "partner/partner"}}{{/embed}}
-    {{/content}}
-
-{{/extend}}
