--- theme/thr/views/partner/partner_loggedin.hbs
@@ -1,15 +1,0 @@
-{{#extend "lib/loggedin" hideSearch="true"}}
-
-    {{#content "scripts" mode="append"}}
-        <script src="{{getAssetPath '/scripts/partners.js'}}" type="text/javascript" nonce="{{nonceValue}}" defer></script>
-    {{/content}}
-
-    {{#content "styles" mode="append"}}
-        <link rel="stylesheet" href="{{getAssetPath '/styles/partner/partner.css'}}"/>
-    {{/content}}
-
-    {{#content "page"}}
-        {{#embed "partner/partner" loggedin="true"}}{{/embed}}
-    {{/content}}
-
-{{/extend}}
