+++ views/template/registration.hbs
@@ -59,8 +59,6 @@
                             <a href="https://hpi.de/impressum.html" target="blank">{{$t "global.link.imprint" }}</a>
                             <span> - </span>
                             <a href="mailto:support@dbildungscloud.de?subject=Schul_Cloud%20Anfrage">{{$t "global.link.contact" }}</a>
-                            <span> - </span>
-                            <a href="https://dbildungscloud.de/partner">{{$t "global.link.partner"}}</a>
                         </nobr>
                     </div>
                 </div>
